separate-artifacts
This change separates the api interfaces from their implementation and creates separate artifacts for each.
