+++ pom.xml
@@ -4,11 +4,15 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>org.codice.usng</groupId>
+    <groupId>org.codice</groupId>
     <artifactId>usng4j</artifactId>
-    <version>0.1</version>
-    <packaging>jar</packaging>
-    <name>Codice USNG</name>
+    <version>0.1-SNAPSHOT</version>
+    <modules>
+        <module>usng4j-api</module>
+        <module>usng4j-impl</module>
+    </modules>
+    <packaging>pom</packaging>
+    <name>usng4j</name>
 
     <properties>
         <junit.version>4.12</junit.version>
@@ -32,17 +36,7 @@
     </scm>
 
     <dependencies>
-        <dependency>
-            <groupId>junit</groupId>
-            <artifactId>junit</artifactId>
-            <version>${junit.version}</version>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.commons</groupId>
-            <artifactId>commons-lang3</artifactId>
-            <version>${commons-lang3.version}</version>
-        </dependency>
+
     </dependencies>
     <build>
         <plugins>
