--- src/main/java/org/codice/usng/BoundingBox.java
@@ -1,58 +1,0 @@
-/* Copyright (c) 2009 Larry Moore, larmoor@gmail.com
- *               2014 Mike Adair, Richard Greenwood, Didier Richard, Stephen Irons, Olivier Terral and Calvin Metcalf (proj4js)
- *               2017 Codice Foundation
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- **/
-
-package org.codice.usng;
-
-/**
- * This interface models an area on the globe represented by a north/south/east/west
- * bounding box.
- *
- * Default implementations of this class are immutable and therefore threadsafe.
- *
- */
-public interface BoundingBox {
-
-    /**
-     *
-     * @return the northern line of latitude for this bounding box.
-     */
-    double getNorth();
-
-    /**
-     *
-     * @return ths southern line of latitude for this bounding box.
-     */
-    double getSouth();
-
-    /**
-     *
-     * @return the eastern line of longitude for this bounding box.
-     */
-    double getEast();
-
-    /**
-     *
-     * @return the western line of longitude for this bounding box.
-     */
-    double getWest();
-}
